refactor `contextual` test files
- moved non-contextual tests into `staged.jl`
- moved `@overlay` tests into `core.jl`
- test `staged.jl` in an interpreter mode
