+++ test/compiler/contextual.jl
@@ -122,127 +122,3 @@ f() = 2
 foo(i) = i+bar(Val(1))
 
 @test @inferred(overdub(Ctx(), foo, 1)) == 43
-
-# overlay method tables
-# =====================
-
-module OverlayModule
-
-using Base.Experimental: @MethodTable, @overlay
-
-@MethodTable(mt)
-
-@overlay mt function sin(x::Float64)
-    1
-end
-
-# short function def
-@overlay mt cos(x::Float64) = 2
-
-# parametric function def
-@overlay mt tan(x::T) where {T} = 3
-
-end # module OverlayModule
-
-let ms = Base._methods_by_ftype(Tuple{typeof(sin), Float64}, nothing, 1, Base.get_world_counter())
-    @test only(ms).method.module === Base.Math
-end
-let ms = Base._methods_by_ftype(Tuple{typeof(sin), Float64}, OverlayModule.mt, 1, Base.get_world_counter())
-    @test only(ms).method.module === OverlayModule
-end
-let ms = Base._methods_by_ftype(Tuple{typeof(sin), Int}, OverlayModule.mt, 1, Base.get_world_counter())
-    @test isempty(ms)
-end
-
-# precompilation
-
-load_path = mktempdir()
-depot_path = mktempdir()
-try
-    pushfirst!(LOAD_PATH, load_path)
-    pushfirst!(DEPOT_PATH, depot_path)
-
-    write(joinpath(load_path, "Foo.jl"),
-          """
-          module Foo
-          Base.Experimental.@MethodTable(mt)
-          Base.Experimental.@overlay mt sin(x::Int) = 1
-          end
-          """)
-
-     # precompiling Foo serializes the overlay method through the `mt` binding in the module
-     Foo = Base.require(Main, :Foo)
-     @test length(Foo.mt) == 1
-
-    write(joinpath(load_path, "Bar.jl"),
-          """
-          module Bar
-          Base.Experimental.@MethodTable(mt)
-          end
-          """)
-
-    write(joinpath(load_path, "Baz.jl"),
-          """
-          module Baz
-          using Bar
-          Base.Experimental.@overlay Bar.mt sin(x::Int) = 1
-          end
-          """)
-
-     # when referring an method table in another module,
-     # the overlay method needs to be discovered explicitly
-     Bar = Base.require(Main, :Bar)
-     @test length(Bar.mt) == 0
-     Baz = Base.require(Main, :Baz)
-     @test length(Bar.mt) == 1
-finally
-    filter!((≠)(load_path), LOAD_PATH)
-    filter!((≠)(depot_path), DEPOT_PATH)
-    rm(load_path, recursive=true, force=true)
-    try
-        rm(depot_path, force=true, recursive=true)
-    catch err
-        @show err
-    end
-end
-
-# Test that writing a bad cassette-style pass gives the expected error (#49715)
-function generator49715(world, source, self, f, tt)
-    tt = tt.parameters[1]
-    sig = Tuple{f, tt.parameters...}
-    mi = Base._which(sig; world)
-
-    error("oh no")
-
-    stub = Core.GeneratedFunctionStub(identity, Core.svec(:methodinstance, :ctx, :x, :f), Core.svec())
-    stub(world, source, :(nothing))
-end
-
-@eval function doit49715(f, tt)
-    $(Expr(:meta, :generated, generator49715))
-    $(Expr(:meta, :generated_only))
-end
-
-@test_throws "oh no" doit49715(sin, Tuple{Int})
-
-# Test that the CodeInfo returned from generated function need not match the
-# generator.
-function overdubbee54341(a, b)
-    a + b
-end
-const overdubee_codeinfo54341 = code_lowered(overdubbee54341, Tuple{Any, Any})[1]
-
-function overdub_generator54341(world::UInt, source::LineNumberNode, args...)
-    if length(args) != 2
-        :(error("Wrong number of arguments"))
-    else
-        return copy(overdubee_codeinfo54341)
-    end
-end
-
-@eval function overdub54341(args...)
-    $(Expr(:meta, :generated, overdub_generator54341))
-    $(Expr(:meta, :generated_only))
-end
-
-@test overdub54341(1, 2) == 3
