+++ test/interpreter.jl
@@ -31,5 +31,8 @@ let p = Pipe(),
     close(p)
 end
 
+# Test generated function behavior in interpreter
+@test success(pipeline(`$(Base.julia_cmd()) --compile=min -E 'include("staged.jl")'`; stderr))
+
 # Test contextual execution mechanism in interpreter (#54360)
 @test success(pipeline(`$(Base.julia_cmd()) --compile=min -E 'include("compiler/contextual.jl")'`; stderr))
